Add `h-get` shell script.
'h-get' is a sort of `apt-get` of Ubuntu package manager but rather just
a small shell utility to help user for preventing to be irritating by
accidental `make` commands. It would be able to replace `Makefile` and
`make` but, not ready for that quality for fow.

If you want to do specific action again, just delete 'state-file' of
your action. (All state-file is resides in root of Haroopress package)
    i.e.:
            rm -f .init.done
            rm -f init-git-submodules.done